app/requirements.txt
requests
python-dotenv
Wand
Flask